#1047 P2 (step 1): session.rs → session/mod.rs + extract key.rs
First step of the structural split called out in #1047. Converts
userspace-dp/src/session.rs into a directory module
userspace-dp/src/session/, and lifts the SessionKey struct plus its
key-transform helpers into session/key.rs.

Going incremental rather than all-in-one PR: this PR extracts the
self-contained key.rs piece (small, well-bounded) without touching
SessionTable, SessionWheel, or the session-entry types. Subsequent
PRs will extract entry.rs / wheel.rs / table.rs from session/mod.rs
once this layout is in place.

What moved into session/key.rs (138 LOC):
- struct SessionKey
- pub(crate) fn forward_wire_key
- pub(crate) fn translated_session_key
- pub(crate) fn reverse_canonical_key
- pub(crate) fn reply_matches_forward_session
- fn reverse_wire_key — visibility widened from file-private `fn` to
  `pub(super) fn` so session/mod.rs's SessionTable impl can still
  call it across the new module boundary

session/mod.rs adds `mod key; pub(crate) use key::*;` re-export so
external callers reaching `crate::session::SessionKey` /
`crate::session::forward_wire_key` etc. continue to resolve.

session_tests.rs renamed to session/tests.rs and the path attribute
on the `mod tests;` declaration in session/mod.rs is updated to match.

LOC distribution after:
- session/mod.rs:   1,234 (was 1,355 in session.rs)
- session/key.rs:     138 (new)
- session/tests.rs: 1,757 (renamed from session_tests.rs, no content change)

cargo test --release: 865 passed, 0 failed.
