+++ userspace-dp/src/session/key.rs
@@ -1,0 +1,138 @@
+// SessionKey + key-transform helpers extracted from session.rs (#1047 P2).
+// Pure relocation — bodies are byte-for-byte identical except
+// `reverse_wire_key` widened from `fn` (file-private) to
+// `pub(super) fn` so session/mod.rs's SessionTable impl can still
+// call it across the new module boundary.
+
+use super::*;
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub(crate) struct SessionKey {
+    pub addr_family: u8,
+    pub protocol: u8,
+    pub src_ip: IpAddr,
+    pub dst_ip: IpAddr,
+    pub src_port: u16,
+    pub dst_port: u16,
+}
+
+pub(crate) fn reply_matches_forward_session(
+    forward_key: &SessionKey,
+    nat: NatDecision,
+    reply_key: &SessionKey,
+) -> bool {
+    reverse_wire_key(forward_key, nat) == *reply_key
+        || reverse_canonical_key(forward_key, nat) == *reply_key
+}
+
+pub(crate) fn forward_wire_key(forward_key: &SessionKey, nat: NatDecision) -> SessionKey {
+    let (src_port, dst_port) = if matches!(forward_key.protocol, PROTO_ICMP | PROTO_ICMPV6) {
+        (forward_key.src_port, forward_key.dst_port)
+    } else {
+        (
+            nat.rewrite_src_port.unwrap_or(forward_key.src_port),
+            nat.rewrite_dst_port.unwrap_or(forward_key.dst_port),
+        )
+    };
+    let wire_src = nat.rewrite_src.unwrap_or(forward_key.src_ip);
+    let wire_dst = nat.rewrite_dst.unwrap_or(forward_key.dst_ip);
+    let (addr_family, protocol) = if nat.nat64 {
+        let af = match wire_src {
+            std::net::IpAddr::V4(_) => libc::AF_INET as u8,
+            std::net::IpAddr::V6(_) => libc::AF_INET6 as u8,
+        };
+        let proto = if af == libc::AF_INET as u8 && forward_key.protocol == PROTO_ICMPV6 {
+            PROTO_ICMP
+        } else if af == libc::AF_INET6 as u8 && forward_key.protocol == PROTO_ICMP {
+            PROTO_ICMPV6
+        } else {
+            forward_key.protocol
+        };
+        (af, proto)
+    } else {
+        (forward_key.addr_family, forward_key.protocol)
+    };
+    SessionKey {
+        addr_family,
+        protocol,
+        src_ip: wire_src,
+        dst_ip: wire_dst,
+        src_port,
+        dst_port,
+    }
+}
+
+pub(crate) fn translated_session_key(key: &SessionKey, nat: NatDecision) -> SessionKey {
+    let (src_port, dst_port) = if matches!(key.protocol, PROTO_ICMP | PROTO_ICMPV6) {
+        (key.src_port, key.dst_port)
+    } else {
+        (
+            nat.rewrite_src_port.unwrap_or(key.src_port),
+            nat.rewrite_dst_port.unwrap_or(key.dst_port),
+        )
+    };
+    SessionKey {
+        addr_family: key.addr_family,
+        protocol: key.protocol,
+        src_ip: nat.rewrite_src.unwrap_or(key.src_ip),
+        dst_ip: nat.rewrite_dst.unwrap_or(key.dst_ip),
+        src_port,
+        dst_port,
+    }
+}
+
+pub(super) fn reverse_wire_key(forward_key: &SessionKey, nat: NatDecision) -> SessionKey {
+    let (src_port, dst_port) = if matches!(forward_key.protocol, PROTO_ICMP | PROTO_ICMPV6) {
+        (forward_key.src_port, forward_key.dst_port)
+    } else {
+        (
+            nat.rewrite_dst_port.unwrap_or(forward_key.dst_port),
+            nat.rewrite_src_port.unwrap_or(forward_key.src_port),
+        )
+    };
+    let wire_src = nat.rewrite_dst.unwrap_or(forward_key.dst_ip);
+    let wire_dst = nat.rewrite_src.unwrap_or(forward_key.src_ip);
+    // NAT64: the reverse (reply) packet is a different address family.
+    // Determine the AF from the NAT-rewritten addresses.
+    let (addr_family, protocol) = if nat.nat64 {
+        let af = match wire_src {
+            std::net::IpAddr::V4(_) => libc::AF_INET as u8,
+            std::net::IpAddr::V6(_) => libc::AF_INET6 as u8,
+        };
+        // ICMPv6 ↔ ICMP protocol mapping.
+        let proto = if af == libc::AF_INET as u8 && forward_key.protocol == PROTO_ICMPV6 {
+            PROTO_ICMP
+        } else if af == libc::AF_INET6 as u8 && forward_key.protocol == PROTO_ICMP {
+            PROTO_ICMPV6
+        } else {
+            forward_key.protocol
+        };
+        (af, proto)
+    } else {
+        (forward_key.addr_family, forward_key.protocol)
+    };
+    SessionKey {
+        addr_family,
+        protocol,
+        src_ip: wire_src,
+        dst_ip: wire_dst,
+        src_port,
+        dst_port,
+    }
+}
+
+pub(crate) fn reverse_canonical_key(forward_key: &SessionKey, _nat: NatDecision) -> SessionKey {
+    let (src_port, dst_port) = if matches!(forward_key.protocol, PROTO_ICMP | PROTO_ICMPV6) {
+        (forward_key.src_port, forward_key.dst_port)
+    } else {
+        (forward_key.dst_port, forward_key.src_port)
+    };
+    SessionKey {
+        addr_family: forward_key.addr_family,
+        protocol: forward_key.protocol,
+        src_ip: forward_key.dst_ip,
+        dst_ip: forward_key.src_ip,
+        src_port,
+        dst_port,
+    }
+}
